# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Application API Documentation",
    "version": "1.0.0",
    "description": "API documentation for the generated application",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "https://your-production-url.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  },
  "security": [],
  "paths": {
...
    "/api/enrollments/{id}": {
      "get": {
        "summary": "Get enrollment by ID",
        "tags": [
          "Enrollments"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "summary": "Update enrollment",
        "tags": [
          "Enrollments"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "summary": "Cancel enrollment",
        "tags": [
          "Enrollments"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/notifications/{id}/read": {
      "put": {
        "summary": "Mark notification as read",
        "tags": [
          "Notifications"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/notifications": {
      "get": {
        "summary": "Get user notifications",
        "tags": [
          "Notifications"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/programs/{id}": {
      "get": {
        "summary": "Get program by ID",
        "tags": [
          "Programs"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "summary": "Update program",
        "tags": [
          "Programs"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "summary": "Delete program",
        "tags": [
          "Programs"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/programs": {
      "get": {
        "summary": "Get all programs",
        "tags": [
          "Programs"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "summary": "Create program",
        "tags": [
          "Programs"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/progress/{employeeId}": {
      "get": {
        "summary": "Get employee learning progress",
        "tags": [
          "Progress"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/progress/{enrollmentId}/update": {
      "put": {
        "summary": "Update progress",
        "tags": [
          "Progress"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/reports/course/{id}": {
      "get": {
        "summary": "Get course analytics",
        "tags": [
          "Reports"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/reports/employee/{id}": {
      "get": {
        "summary": "Get employee learning report",
        "tags": [
          "Reports"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
...
}